LayoutTests/js/arrowfunction-associativity-1-expected.txt

Tests for ES6 arrow function nested declaration

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


af1 = af2 = af3 => af1 = af2 = af3
PASS af1 is af2
PASS af2 is 13
PASS af1 is 13
PASS successfullyParsed is true

TEST COMPLETE
